Fix a panic in diff display when parsing YAML strings
To make sure all of the string was consumed by the YAML decoder, we'd call `yamlDecoder.Decode(nil)` to see if it returned an `io.EOF` error. However, if it's not EOF, the decoder panics when `nil` is passed-in. Instead, pass in a valid argument for cases that aren't EOF.