feat: enable imports from nodejs provider resources (#14668)
# Description
I was writing a typescript resource provider, but got errors that the
the plugin didn't support imports, while the `read` method was
implemented on the provider.
This PR also sets the `Inputs` on the `ReadResponse`, so imports will be
supported.

Fixes # (issue)
Could not find existing issue

## Checklist

- [ ] I have run `make tidy` to update any new dependencies
- [ ] I have run `make lint` to verify my code passes the lint check
- [ ] I have formatted my code using `gofumpt`
(None of the above, since the change is in js code)

- [ ] I have added tests that prove my fix is effective or that my
feature works
There are no existing tests for the `server.ts` file. If really required
for merging I could maybe take a look at a test for this use-case.

- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change

- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version