Fix: Don't delete stack outputs on failed deployments
If an update fails, update any stack outputs that were updated, but otherwise leave existing stack outputs as-is. In other words, don't delete stack outputs if the stack didn't successfully run to completion.

Implementation note: when we receive a RegisterResourceOutputsEvent for the stack resource, we defer processing it until the end of the deployment, to know whether the deployment ran to completion without errors.