Sets the `main` project setting to the cwd for inline programs
Currently if no project settings are provided this is the default value for
`main`, but if the caller providers a project_settings object without this set
it will cause a module load issue if the inline program depends on other local
python files. The current work-around is to put all the code in one file or set
`work_dir`.

also fix a test warning by renaming `test_path` to `get_test_path`

Fixes isssue #7960