[sdk-gen/go] Fix compiling plain element type with plain maps (#14704)
# Description

Fixes #14662 when choosing whether to use `ObjectName` (plain) or
`ObjectNameArgs` (non-plain) we have to check whether the non-plain
version was actually available in first place (sometimes it isn't when
the type isn't used anywhere as non-plain) so the fix here is to use the
package context that tracks the usages to determine whether the `Args`
suffix should be used.

Funny though, for other languages like C#, TypeScript and Python we
don't have a concept of a strictly plain type. Instead all types are
non-plain (accepting values for their properties being inputs/outputs)
which means that a plain element type on maps for these languages has no
effect

## Checklist

- [ ] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [x] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [x] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->