Just call plugin install directly in tests (#14723)
<!--- 
Thanks so much for your contribution! If this is your first time
contributing, please ensure that you have read the
[CONTRIBUTING](https://github.com/pulumi/pulumi/blob/master/CONTRIBUTING.md)
documentation.
-->

# Description

<!--- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context. -->

Pointed out in another PR
https://github.com/pulumi/pulumi/pull/14548#discussion_r1411431491.

No need to check if a plugin is already installed with `plugin ls`,
`install` will return success if the plugin is already present.