[ci] Fix Run @actions/setup-go flake (#14798)
Bumps @actions/setup-go to v5.

.github/workflows/ci-run-test.yml uses @actions/setup-go@v3 which uses
@actions/cache@v3.

This version suffers from a bug where it will hang forever when
downloading a segment and will timeout the job.

According to
https://github.com/actions/cache/blob/6a1a45d/tips-and-workarounds.md#cache-segment-restore-timeout
v3.0.8 adds SEGMENT_DOWNLOAD_TIMEOUT_MINS with a default of 10 minutes
which should alleviate our CI hangs.

Only applying this to `ci-run-test.yml` to limit the scope of the
change.