[test] fix `TestCreatingProjectWithEmptyConfig` flakiness (#14777)
Fixes #14756

Flaky tests cause coverage to be rerun and drops coverage data making
CodeCov data less reliable/usable.

One of the most commonly flaky tests is
`TestCreatingProjectWithEmptyConfig` in `pkg` which contains a large
number of tests.

Moving this test to the top of the file appears to lead to less
flakiness.