use uuid instead of random number for test suffix
Currently we're using a random number for getting the test suffix.
However given the birthday paradox and the number of tests we're
running this doesn't seem to be good enough to keep our tests from
being flaky.  Use a uuid instead, which is virtually guaranteed to be
unique, and should eliminate this problem.