[ci] Re-enable code coverage and rebuild release binaries without coverage (#14806)
Fixes #14817 

As a follow-up to #14804, we need to re-enable code coverage without
releases being built with coverage. #14804 includes a test in the
release job to catch `pulumi` being released with coverage data.

Also fixes the test to check that the pulumi binary is not built with
coverage.

Background:
Coverage was disabled on merge jobs in order to fix
[pulumi/pulumi#14799].
[pulumi/pulumi#14716] removed nightly coverage workflows in favor of
collecting coverage on merges [pulumi/pulumi#14727].
The binaries built with coverage were used in the downstream
prepare-release job causing a noticable bug [pulumi/pulumi#14799].