increase the timeout for automation API tests
These tests are pretty close to hitting the timeout during regular
runs (e.g. one successful run I picked randomly took ~194000ms).  The
GitHub action runners don't have super reliable performance, so we can
easily get pushed over this limit.

To make matters worse here, if we hit this timeout just at the right
time, the test doesn't exit cleanly (potentially related to
https://github.com/pulumi/pulumi-dotnet/issues/134), as I've seen a
`pulumi preview` process stick around in that case, preventing the
tests from shutting down).  This means we have to wait until the CI
job times out after an hour until the failure is reported.

I'm not sure if we only got close to this timeout recently, or if this
is a longer standing issue, but it reproduces well for me locally.
Ideally of course the tests would be faster, but this is still "only"
~5min which is much faster than other tests, and should hopefully
reduce the amount of times we need to go through the merge queue,
saving a lot of time there.