langhost test: retry if the langhost is unavailable (#14843)
This test is racy, as the langhost can sometimes take a little bit too
long to be ready to accept connections successfully. Just retry a few
times in this case, to make sure it's ready to have the test run against
it.

I haven't been able to reliable reproduce the flakyness locally, so
there's no guarantee that this will actually do what it says
successfully. I think it's still worth a try and we can monitor CI to
verify if this fixed the issue, and revert this PR if the same issue
appears again.

Fixes https://github.com/pulumi/pulumi/issues/14841