Add LogicalName to import system (#14745)
<!--- 
Thanks so much for your contribution! If this is your first time
contributing, please ensure that you have read the
[CONTRIBUTING](https://github.com/pulumi/pulumi/blob/master/CONTRIBUTING.md)
documentation.
-->

# Description

<!--- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context. -->

Fixes https://github.com/pulumi/pulumi/issues/14618.

This adds a new field to the import file "logicalName", which if set is
used as the logical name
(https://www.pulumi.com/docs/concepts/resources/names/#logicalname) of
the resource. The original "name" field is then used just for codegen
purposes (i.e. the source name).

If either field is not set then the other field is used to fill it in.
If neither field is set it's an error.

Further the import system _no longer_ renames resources to try and make
unique names, but it also no longer errors just because two resources of
different types have the same name. The rules for uniqueness now match
what's valid when writing a Pulumi program, but it is the users
responsibility to make names unique and the import system will error if
you try to import two resources that would end up with the same URN.

## Checklist

- [x] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [x] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [x] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->