correct version check when we have a dev version installed
When we have a dev version from the dev CLI channel installed, we want
to warn whenever a new version is available, so the users can upgrade.
Do that, and also reduce the amount of time the version information is
cached, since we expect dev versions to be released much more
frequently than regular releases.