ci-dev-release: remove extra quotes (#14964)
GitHub actions interprets these quotes literally, so the version we end
up with includes the quotes, and will therefore end up in the form of
`v"1.0.0-11-gdeadbeef" in the output. Remove the extra quotes as they
are not needed.

It's a bit awkward to test this, but I *think* this should do the trick.